requirements.txt
glfw==2.5.1
imgui==1.4.1
numpy==1.22.3
opencv-python==4.5.5.64
PyOpenGL==3.1.6
